[feat] update

+++ src/main/java/com/tave/gangnam/week3/assigment/dto/LoanRequestDto.java
@@ -15,5 +15,5 @@ public class LoanRequestDto {
 
     private String loanName;
 
-    private BigDecimal loanAmount;
+    private Long loanAmount;
 }
